히트맵 정상작동 및 SpringConfig수정

+++ backend/src/main/java/lx/edu/levelup/feature/heatmap/HeatMapDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 /**
  * 히트맵 데이터 전송 객체 (Data Transfer Object)
  * Controller가 프론트엔드와 데이터를 주고받을 때 사용합니다.
@@ -12,7 +14,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor // 파라미터 없는 기본 생성자
 @AllArgsConstructor // 모든 필드를 포함하는 생성자
 public class HeatMapDTO {
-    private double lat;
-    private double lng;
+    private BigDecimal lat;
+    private BigDecimal lng;
     
 }
